+++ packages/@pollyjs/ember/blueprints/@pollyjs/ember/files/config/polly.js
@@ -1,0 +1,13 @@
+/* eslint-env node */
+
+'use strict';
+
+module.exports = function(env) {
+  return {
+    enabled: env !== 'production',
+    server: {
+      apiNamespace: '/polly',
+      recordingsDir: 'recordings'
+    }
+  };
+};