+++ packages/@pollyjs/ember/index.js
@@ -2,18 +2,49 @@
 
 'use strict';
 
+const fs = require('fs');
 const path = require('path');
 const { registerExpressAPI, Defaults } = require('@pollyjs/node-server');
+const parseArgs = require('minimist');
 
-const { assign } = Object;
+function determineEnv() {
+  if (process.env.EMBER_ENV) {
+    return process.env.EMBER_ENV;
+  }
+
+  let args = parseArgs(process.argv);
+  let env = args.e || args.env || args.environment;
+
+  // Is it "ember b -prod" or "ember build --prod" command?
+  if (
+    !env &&
+    (process.argv.indexOf('-prod') > -1 || process.argv.indexOf('--prod') > -1)
+  ) {
+    env = 'production';
+  }
+
+  // Is it "ember test" or "ember t" command without explicit env specified?
+  if (
+    !env &&
+    (process.argv.indexOf('test') > -1 || process.argv.indexOf('t') > -1)
+  ) {
+    env = 'test';
+  }
+
+  return env || 'development';
+}
 
 module.exports = {
   name: require('./package').name,
+  _config: null,
 
-  included() {
-    this._super.included.apply(this, arguments);
+  init() {
+    // see: https://github.com/ember-cli/ember-cli/blob/725e129e62bccbf21af55b21180edb8966781f53/lib/models/addon.js#L258
+    this._super.init && this._super.init.apply(this, arguments);
 
-    this._config = this._pollyConfig();
+    const env = determineEnv();
+
+    this._config = this._pollyConfig(env);
   },
 
   treeForAddon() {
@@ -37,6 +68,29 @@ module.exports = {
     `;
   },
 
+  _pollyConfig(env) {
+    // defaults
+    let config = {
+      enabled: env !== 'production',
+      server: {}
+    };
+
+    let configPath = path.join(this.root, 'config', 'polly.js');
+
+    if (fs.existsSync(configPath)) {
+      let configGenerator = require(configPath);
+
+      Object.assign(config, configGenerator(env));
+    }
+
+    config.server.recordingsDir = path.join(
+      this.root,
+      config.server.recordingsDir || Defaults.recordingsDir
+    );
+
+    return config;
+  },
+
   serverMiddleware(startOptions) {
     this.testemMiddleware(startOptions.app);
   },
@@ -45,22 +99,5 @@ module.exports = {
     if (this._config.enabled) {
       registerExpressAPI(app, this._config.server);
     }
-  },
-
-  _pollyConfig() {
-    const config = assign(
-      {
-        enabled: this.app.env !== 'production',
-        server: {}
-      },
-      this.app.options.pollyjs
-    );
-
-    config.server.recordingsDir = path.join(
-      this.app.project.root,
-      config.server.recordingsDir || Defaults.recordingsDir
-    );
-
-    return config;
   }
 };